+++ src/main/java/com/fondant/product/presentation/dto/response/ProductDetailResponse.java
@@ -16,6 +16,6 @@ public record ProductDetailResponse(
         String description,
         List<ImageInfo> detailPages,
         MarketInfoForProductDetail marketInfo,
-        int basePrice
+        Double basePrice
 ) {
 }
